Issue 77 - Fixed checkstyle issues.
